Kirill_Madorin/requirements.txt
pandas>=2.0.0
nltk>=3.8.0
openai>=1.18.0
openai-agents>=0.0.9
argparse>=1.4.0
tqdm>=4.65.0
python-dotenv>=1.0.0 